WIP: split trip editor place mutation service (checkpoint)
